FIX: 수정버전을 기존의 버전으로 이름 변경 및 기존 API 삭제, 테스트 수정
